Remove X check on GetSpawnPoint
The SpawnX >= 0 check on GetSpawnPoint will erroneously cause GetSpawnPoint to return the player's actual spawn point only if it is at world X center or to the east of center. It should return the spawn location even if east of center. It is understandable, though, why you might not want it to be underground, so I kept the SpawnY >= 0 check.